Job 3 & sort OK (TF-IDF)

+++ src/main/java/org/formation/hadoop/WordFilteringApp.java
@@ -3,6 +3,7 @@ package org.formation.hadoop;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
@@ -20,7 +21,7 @@ public class WordFilteringApp {
     private static final URI stopwordsURI = new Path("cache/TF-IDF/stopwords_en.txt").toUri();
 
     public static void main(String[] args) {
-        int startFrom = 0;
+        int startFrom = 0, endTask = 10;
         if (args.length < 2) {
             usage();
             System.exit(1);
@@ -28,6 +29,10 @@ public class WordFilteringApp {
         if (args.length > 2) {
             startFrom = Integer.parseInt(args[2]);
         }
+        if (args.length > 3) {
+            endTask = Integer.parseInt(args[3]);
+        }
+
         try {
             Configuration configuration = new Configuration();
             FileSystem fileSystem = FileSystem.newInstance(configuration);
@@ -45,8 +50,8 @@ public class WordFilteringApp {
                 fileSystem.delete(outputFilePath);
             }
 
-            if(startFrom < 1) { // First job (filter and word count):
-
+            if(startFrom < 1 && endTask >= 0) { // First job (filter and word count):
+                System.out.println("Starting Job 1");
                 Job job = new Job(configuration);
                 if(fileSystem.exists(tmpOutputFile[0])) {
                     fileSystem.delete(tmpOutputFile[0]);
@@ -70,12 +75,16 @@ public class WordFilteringApp {
                 job.waitForCompletion(true);
             }
 
-            if (startFrom < 2){ // Seccond job (count word per doc):
+            if (startFrom < 2 && endTask >= 1) { // Seccond job (count word per doc):
+                System.out.println("Starting Job 2");
                 Job job = new Job(configuration);
                 job.setJobName("Word Total Count");
+                if(fileSystem.exists(tmpOutputFile[1])) {
+                    fileSystem.delete(tmpOutputFile[1]);
+                }
 
                 FileInputFormat.addInputPath(job, tmpOutputFile[0]);
-                FileOutputFormat.setOutputPath(job, outputFilePath);
+                FileOutputFormat.setOutputPath(job, tmpOutputFile[1]);
 
                 job.setOutputKeyClass(CustomKey.class);
                 job.setOutputValueClass(CustomValue.class);
@@ -85,6 +94,49 @@ public class WordFilteringApp {
                 job.setGroupingComparatorClass(org.formation.hadoop.jobs.wordPerDoc.GroupingComparator.class);
                 job.setReducerClass(org.formation.hadoop.jobs.wordPerDoc.Reduce.class);
 
+                job.setJarByClass(WordFilteringApp.class);
+                job.setNumReduceTasks(1);
+
+                job.waitForCompletion(true);
+            }
+
+            if (startFrom < 3 && endTask >= 2) { // Last job (TF-IDF):
+                System.out.println("Starting Job 3");
+                Job job = new Job(configuration);
+                job.setJobName("TF-IDF computation");
+                if(fileSystem.exists(tmpOutputFile[2])) {
+                    fileSystem.delete(tmpOutputFile[2]);
+                }
+
+                FileInputFormat.addInputPath(job, tmpOutputFile[1]);
+                FileOutputFormat.setOutputPath(job, tmpOutputFile[2]);
+
+                job.setOutputKeyClass(CustomKey.class);
+                job.setMapOutputValueClass(CustomValue.class);
+                job.setOutputValueClass(DoubleWritable.class);
+
+                job.setMapperClass(org.formation.hadoop.jobs.computeTFIDF.Map.class);
+                job.setPartitionerClass(org.formation.hadoop.jobs.computeTFIDF.Partitioner.class);
+                job.setGroupingComparatorClass(org.formation.hadoop.jobs.computeTFIDF.GroupingComparator.class);
+                job.setReducerClass(org.formation.hadoop.jobs.computeTFIDF.Reduce.class);
+
+                job.setJarByClass(WordFilteringApp.class);
+
+                job.waitForCompletion(true);
+            }
+
+            if (startFrom < 4&& endTask >= 3){ // sorting on TF-IDF values:
+                System.out.println("Starting Job 4");
+                Job job = new Job(configuration);
+                job.setJobName("TF-IDF sorting");
+
+                FileInputFormat.addInputPath(job, tmpOutputFile[2]);
+                FileOutputFormat.setOutputPath(job, outputFilePath);
+
+                job.setOutputKeyClass(DoubleWritable.class);
+                job.setOutputValueClass(CustomKey.class);
+
+                job.setMapperClass(org.formation.hadoop.jobs.sort.Map.class);
                 job.setJarByClass(WordFilteringApp.class);
 
                 job.waitForCompletion(true);
